Renamed some file names related to the pools and deleted the empty resource_pool files.
