+++ src/constant_buffer_pool.cpp
@@ -1,4 +1,4 @@
-#include "resource_pool_constant_buffer.hpp"
+#include "constant_buffer_pool.hpp"
 
 namespace wr 
 {
