+++ src/d3d12/d3d12_model_pool.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../resource_pool_model.hpp"
+#include "../model_pool.hpp"
 #include "d3d12_structs.hpp"
 
 #include <queue>
