+++ src/d3d12/d3d12_renderer.cpp
@@ -7,10 +7,10 @@
 #include "../window.hpp"
 
 #include "d3d12_defines.hpp"
-#include "d3d12_resource_pool_material.hpp"
-#include "d3d12_resource_pool_model.hpp"
-#include "d3d12_resource_pool_constant_buffer.hpp"
-#include "d3d12_resource_pool_structured_buffer.hpp"
+#include "d3d12_material_pool.hpp"
+#include "d3d12_model_pool.hpp"
+#include "d3d12_constant_buffer_pool.hpp"
+#include "d3d12_structured_buffer_pool.hpp"
 #include "d3d12_functions.hpp"
 #include "d3d12_pipeline_registry.hpp"
 #include "d3d12_shader_registry.hpp"
