+++ src/material_pool.cpp
@@ -1,6 +1,6 @@
 #define NOMINMAX
 
-#include "resource_pool_material.hpp"
+#include "material_pool.hpp"
 
 #include <assimp/material.h>
