+++ src/scene_graph/scene_graph.cpp
@@ -12,8 +12,7 @@
 #include "../d3d12/d3d12_defines.hpp"
 #include "../d3d12/d3d12_functions.hpp"
 #include "../d3d12/d3d12_renderer.hpp"
-#include "../d3d12/d3d12_resource_pool.hpp"
-#include "../d3d12/d3d12_resource_pool_constant_buffer.hpp"
+#include "../d3d12/d3d12_constant_buffer_pool.hpp"
 
 namespace wr
 {
